FIX: GEN dataset now returns DoubleTensor

+++ src/datasets.py
@@ -434,7 +434,7 @@ class GENdatasetILC(Dataset):
             rx, ry = int(proj.shape[0] / 2), int(proj.shape[1] / 2)
             image = image[:, mx - rx:mx + rx, my - ry:my + ry]
             image = local_autoscale_ms(image)
-            return image, label
+            return image.float(), label
 
         if np.random.randint(0, 2) == 0:
             pil_img = Image.fromarray(image[:, :, 0])
